Create an account app for the website.
Create an app named "account" for the website that is used to work with users and sessions.
Create a login form, this form authenticates users to the database.
Add the buil-in Django authentication log in and log out views, configure url patterns for them.
Create templates for both views and define a simple dashboard view to redirect users after they log in.
Configure the settings file to use the LOG_IN and LOG_OUT URLs by default.