Add user authentication using Facebook and Google. Create an images app.
Add social authentication backends for Facebook and Google to allow
users to authenticate using their social media accounts.
Create an Image app, that will allow users to bookmark and share
images they find on other websites or on this site.
Define a model to store images and their information called "Image".
Create a form and a view to handle image uploads. Configure a url
path and create a template for the view.