Create a detail view for pictures and add an easy-thumbnails app.
Create a detail view to display images that have been bookmarked on the site.
Add a url pattern for this view and define a get_absolute_url method for the
Image model. Create a detail.html template for the view.
Add a django app called easy-thumbnails. Thumbnails will load faster in the
browser and homogenize images of very different size. Add a thumbnail
template tag to the detail.html template.